chore: Remove CLI.
I would like to build a better CLI with multiple subcommands but there is a "limitation"
in `commander` that would require me to split my CLI into multiple binaries. I've decided
that the API is easy enough to not need a CLI for local testing and it's better to keep
this package focused on delivering the library.